+++ vcfpy/record.py
@@ -148,6 +148,19 @@ class Record:
         """Return generator yielding from ``self.calls``"""
         yield from self.calls
 
+    def __eq__(self, other):
+        if isinstance(other, self.__class__):
+            return self.__dict__ == other.__dict__
+        return NotImplemented
+
+    def __ne__(self, other):
+        if isinstance(other, self.__class__):
+            return not self.__eq__(other)
+        return NotImplemented
+
+    def __hash__(self):
+        return hash(tuple(sorted(self.__dict__.items())))
+
     def __str__(self):
         tpl = 'Record({})'
         lst = [self.CHROM, self.POS, self.ID, self.REF, self.ALT, self.QUAL,
@@ -261,6 +274,19 @@ class Call:
         """Return ``True`` for non-hom-ref calls"""
         return self.gt_type != HOM_REF
 
+    def __eq__(self, other):
+        if isinstance(other, self.__class__):
+            return self.__dict__ == other.__dict__
+        return NotImplemented
+
+    def __ne__(self, other):
+        if isinstance(other, self.__class__):
+            return not self.__eq__(other)
+        return NotImplemented
+
+    def __hash__(self):
+        return hash(tuple(sorted(self.__dict__.items())))
+
     def __str__(self):
         tpl = 'Call({})'
         lst = [self.sample, self.data]
@@ -282,6 +308,19 @@ class AltRecord:
         #: header lines, such as DUP, DEL, INS, ...
         self.type = type_
 
+    def __eq__(self, other):
+        if isinstance(other, self.__class__):
+            return self.__dict__ == other.__dict__
+        return NotImplemented
+
+    def __ne__(self, other):
+        if isinstance(other, self.__class__):
+            return not self.__eq__(other)
+        return NotImplemented
+
+    def __hash__(self):
+        return hash(tuple(sorted(self.__dict__.items())))
+
     def serialize(self):
         """Return ``str`` with representation for VCF file"""
         raise NotImplementedError(
@@ -302,6 +341,19 @@ class Substitution(AltRecord):
 
     def serialize(self):
         return self.value
+
+    def __eq__(self, other):
+        if isinstance(other, self.__class__):
+            return self.__dict__ == other.__dict__
+        return NotImplemented
+
+    def __ne__(self, other):
+        if isinstance(other, self.__class__):
+            return not self.__eq__(other)
+        return NotImplemented
+
+    def __hash__(self):
+        return hash(tuple(sorted(self.__dict__.items())))
 
     def __str__(self):
         tpl = 'Substitution(type_={}, value={})'
@@ -359,6 +411,19 @@ class BreakEnd(AltRecord):
         else:
             return self.sequence + remote_tag
 
+    def __eq__(self, other):
+        if isinstance(other, self.__class__):
+            return self.__dict__ == other.__dict__
+        return NotImplemented
+
+    def __ne__(self, other):
+        if isinstance(other, self.__class__):
+            return not self.__eq__(other)
+        return NotImplemented
+
+    def __hash__(self):
+        return hash(tuple(sorted(self.__dict__.items())))
+
     def __str__(self):
         tpl = 'BreakEnd({})'
         vals = [
@@ -375,6 +440,19 @@ class SingleBreakEnd(BreakEnd):
 
     def __init__(self, orientation, sequence):
         super().__init__(None, None, orientation, None, sequence, None)
+
+    def __eq__(self, other):
+        if isinstance(other, self.__class__):
+            return self.__dict__ == other.__dict__
+        return NotImplemented
+
+    def __ne__(self, other):
+        if isinstance(other, self.__class__):
+            return not self.__eq__(other)
+        return NotImplemented
+
+    def __hash__(self):
+        return hash(tuple(sorted(self.__dict__.items())))
 
     def __str__(self):
         tpl = 'SingleBreakEnd({})'
@@ -397,6 +475,19 @@ class SymbolicAllele(AltRecord):
 
     def serialize(self):
         return '<{}>'.format(self.value)
+
+    def __eq__(self, other):
+        if isinstance(other, self.__class__):
+            return self.__dict__ == other.__dict__
+        return NotImplemented
+
+    def __ne__(self, other):
+        if isinstance(other, self.__class__):
+            return not self.__eq__(other)
+        return NotImplemented
+
+    def __hash__(self):
+        return hash(tuple(sorted(self.__dict__.items())))
 
     def __str__(self):
         return 'SymbolicAllele({})'.format(repr(self.value))